+++ src/NxT.Core/Models/SalesRecord.cs
@@ -11,13 +11,13 @@ public class SalesRecord
     public DateTime Date { get; set; }
 
     [DisplayFormat(DataFormatString = "{0:C2}")]
-    public double Amount { get; set; }
+    public decimal Amount { get; set; }
     public ESaleStatus Status { get; set; }
     public int SellerID { get; set; }
     public Seller Seller { get; set; } = null!;
 
     public SalesRecord() { }
-    public SalesRecord(int id, DateTime date, double amount, ESaleStatus status, Seller seller)
+    public SalesRecord(int id, DateTime date, decimal amount, ESaleStatus status, Seller seller)
     {
         ID = id;
         Date = date;
